refactor: extract user controls into separate component
